Migrate AllowsColourOutput to cli-runtime printers package
